Day15 + PriorityQueue
FossilOrigin-Name: a48dd08ee51e71a48792226940bf22b30ebd44b6fc97c745e27482a0091c273d

+++ day11/day11.go
@@ -49,13 +49,13 @@ func updateState(state [][]int, updateFn func(x, y int, cell *int)) {
 	}
 }
 
-func incState(state [][]int, pairs map[Pair]int) {
+func incState(state [][]int, pairs map[Pair]int64) {
 	updateState(state, func(x, y int, cell *int) {
 		var inc = pairs[Pair{x, y}]
 		if pairs == nil {
 			inc = 1
 		}
-		*cell += inc
+		*cell += int(inc)
 	})
 }
